@@ -101,8 +101,7 @@ public class Migrator
         runSettingsXmlDoc.Load(reader);
         var root = runSettingsXmlDoc.DocumentElement;
 
-        var testSettingsNode = root.SelectSingleNode(@"/RunSettings/MSTest/SettingsFile");
-
+        var testSettingsNode = root?.SelectSingleNode(@"/RunSettings/MSTest/SettingsFile");
         if (testSettingsNode != null)
         {
             testSettingsPath = testSettingsNode.InnerText;
@@ -113,7 +112,7 @@ public class Migrator
             // Expand path relative to runSettings location.
             if (!Path.IsPathRooted(testSettingsPath))
             {
-                testSettingsPath = Path.GetFullPath(Path.Combine(Path.GetDirectoryName(oldRunSettingsPath), testSettingsPath));
+                testSettingsPath = Path.GetFullPath(Path.Combine(Path.GetDirectoryName(oldRunSettingsPath)!, testSettingsPath));
             }
 
             // Remove the embedded testSettings node if it exists.
@@ -153,36 +152,42 @@ public class Migrator
     /// <param name="runSettingsXmlDoc">Runsettings Xml</param>
     private static void MigrateTestSettingsNodesToRunSettings(string testSettingsPath, XmlDocument runSettingsXmlDoc)
     {
-        var testSettingsNodes = ReadTestSettingsNodes(testSettingsPath);
+        TestSettingsNodes testSettingsNodes = ReadTestSettingsNodes(testSettingsPath);
 
         string? testTimeout = null;
-        if (testSettingsNodes.Timeout != null && testSettingsNodes.Timeout.Attributes[TestTimeoutAttributeName] != null)
-        {
-            testTimeout = testSettingsNodes.Timeout.Attributes[TestTimeoutAttributeName].Value;
-        }
-
         string? runTimeout = null;
-        if (testSettingsNodes.Timeout != null && testSettingsNodes.Timeout.Attributes[RunTimeoutAttributeName] != null)
+        if (testSettingsNodes.Timeout is not null)
         {
-            runTimeout = testSettingsNodes.Timeout.Attributes[RunTimeoutAttributeName].Value;
+            if (testSettingsNodes.Timeout.Attributes?[TestTimeoutAttributeName] is XmlAttribute testTimeoutAttribute)
+            {
+                testTimeout = testTimeoutAttribute.Value;
+            }
+
+            if (testSettingsNodes.Timeout.Attributes?[RunTimeoutAttributeName] is XmlAttribute runTimeoutAttribute)
+            {
+                runTimeout = runTimeoutAttribute.Value;
+            }
         }
 
         string? parallelTestCount = null;
-        if (testSettingsNodes.Execution != null && testSettingsNodes.Execution.Attributes[ParallelTestCountAttributeName] != null)
-        {
-            parallelTestCount = testSettingsNodes.Execution.Attributes[ParallelTestCountAttributeName].Value;
-        }
-
         string? hostProcessPlatform = null;
-        if (testSettingsNodes.Execution != null && testSettingsNodes.Execution.Attributes[HostProcessPlatformAttributeName] != null)
+        if (testSettingsNodes.Execution is not null)
         {
-            hostProcessPlatform = testSettingsNodes.Execution.Attributes[HostProcessPlatformAttributeName].Value;
+            if (testSettingsNodes.Execution.Attributes?[ParallelTestCountAttributeName] is XmlAttribute parallelTestCountAttribute)
+            {
+                parallelTestCount = parallelTestCountAttribute.Value;
+            }
+
+            if (testSettingsNodes.Execution.Attributes?[HostProcessPlatformAttributeName] is XmlAttribute hostProcessPlatformAttribute)
+            {
+                hostProcessPlatform = hostProcessPlatformAttribute.Value;
+            }
         }
 
         // WebTestRunConfiguration node.
         if (testSettingsNodes.WebSettings != null)
         {
-            runSettingsXmlDoc.DocumentElement.AppendChild(runSettingsXmlDoc.ImportNode(testSettingsNodes.WebSettings, deep: true));
+            runSettingsXmlDoc.DocumentElement?.AppendChild(runSettingsXmlDoc.ImportNode(testSettingsNodes.WebSettings, deep: true));
         }
 
         // LegacySettings node.
@@ -191,11 +196,11 @@ public class Migrator
         // TestSessionTimeout node.
         if (!runTimeout.IsNullOrEmpty())
         {
-            AddRunTimeoutNode(runTimeout, runSettingsXmlDoc);
+            AddRunTimeoutNode(runTimeout, runSettingsXmlDoc!);
         }
 
         // DataCollectors node.
-        if (testSettingsNodes.Datacollectors != null && testSettingsNodes.Datacollectors.Count > 0)
+        if (testSettingsNodes?.Datacollectors != null && testSettingsNodes.Datacollectors.Count > 0)
         {
             AddDataCollectorNodes(testSettingsNodes.Datacollectors, runSettingsXmlDoc);
         }
@@ -203,7 +208,7 @@ public class Migrator
 
     private static TestSettingsNodes ReadTestSettingsNodes(string testSettingsPath)
     {
-        var testSettingsNodes = new TestSettingsNodes();
+        TestSettingsNodes testSettingsNodes = new();
 
         using (XmlTextReader reader = new(testSettingsPath))
         {
@@ -211,7 +216,7 @@ public class Migrator
 
             var testSettingsXmlDoc = new XmlDocument();
             testSettingsXmlDoc.Load(reader);
-            var testSettingsRoot = testSettingsXmlDoc.DocumentElement;
+            var testSettingsRoot = testSettingsXmlDoc.DocumentElement!;
 
             // Select the interesting nodes from the xml.
             testSettingsNodes.Deployment = testSettingsRoot.SelectSingleNode(@"/TestSettings/Deployment");
@@ -223,14 +228,16 @@ public class Migrator
             testSettingsNodes.Hosts = testSettingsRoot.SelectSingleNode(@"/TestSettings/Execution/Hosts");
             testSettingsNodes.Execution = testSettingsRoot.SelectSingleNode(@"/TestSettings/Execution");
 
-            if (testSettingsNodes.Timeout != null && (testSettingsNodes.Timeout.Attributes[AgentNotRespondingTimeoutAttribute] != null ||
-                                                      testSettingsNodes.Timeout.Attributes[DeploymentTimeoutAttribute] != null || testSettingsNodes.Timeout.Attributes[ScriptTimeoutAttribute] != null))
+            if (testSettingsNodes.Timeout != null
+                && (testSettingsNodes.Timeout.Attributes?[AgentNotRespondingTimeoutAttribute] != null
+                    || testSettingsNodes.Timeout.Attributes?[DeploymentTimeoutAttribute] != null
+                    || testSettingsNodes.Timeout.Attributes?[ScriptTimeoutAttribute] != null))
             {
                 Console.WriteLine(CommandLineResources.UnsupportedAttributes);
             }
         }
 
-        return testSettingsNodes;
+        return testSettingsNodes!;
     }
 
     /// <summary>
@@ -239,10 +246,10 @@ public class Migrator
     /// <param name="newXmlDoc">Xml doc to process</param>
     private static void RemoveEmbeddedTestSettings(XmlDocument newXmlDoc)
     {
-        var testSettingsNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/MSTest/SettingsFile");
+        var testSettingsNode = newXmlDoc.DocumentElement?.SelectSingleNode(@"/RunSettings/MSTest/SettingsFile");
         if (testSettingsNode != null)
         {
-            testSettingsNode.ParentNode.RemoveChild(testSettingsNode);
+            testSettingsNode.ParentNode?.RemoveChild(testSettingsNode);
         }
     }
 
@@ -267,9 +274,13 @@ public class Migrator
             return;
         }
 
+        if (newXmlDoc.DocumentElement == null)
+        {
+            return;
+        }
+
         // Add ForcedLegacy node.
         var mstestNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/MSTest");
-        XmlNode forcedLegacyNode;
         if (mstestNode == null)
         {
             mstestNode = newXmlDoc.CreateNode(XmlNodeType.Element, MSTestNodeName, null);
@@ -277,19 +288,24 @@ public class Migrator
             mstestNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/MSTest");
         }
 
-        forcedLegacyNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/MSTest/ForcedLegacyMode");
+        if (mstestNode == null)
+        {
+            return;
+        }
+
+        var forcedLegacyNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/MSTest/ForcedLegacyMode");
         if (forcedLegacyNode == null)
         {
             forcedLegacyNode = newXmlDoc.CreateNode(XmlNodeType.Element, ForcedLegacyModeName, null);
             mstestNode.AppendChild(newXmlDoc.ImportNode(forcedLegacyNode, deep: true));
-            forcedLegacyNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/MSTest/ForcedLegacyMode");
+            forcedLegacyNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/MSTest/ForcedLegacyMode")!;
         }
 
         forcedLegacyNode.InnerText = "true";
 
         // Remove if the legacy node already exists.
         var legacyNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/LegacySettings");
-        if (legacyNode != null)
+        if (legacyNode?.ParentNode != null)
         {
             Console.WriteLine(CommandLineResources.IgnoringLegacySettings);
             legacyNode.ParentNode.RemoveChild(legacyNode);
@@ -310,28 +326,28 @@ public class Migrator
         // Execution node.
         if (testSettingsNodes.UnitTestConfig != null || !parallelTestCount.IsNullOrEmpty() || !testTimeout.IsNullOrEmpty() || testSettingsNodes.Hosts != null)
         {
-            var newExecutionNode = newXmlDoc.CreateNode(XmlNodeType.Element, ExecutionNodeName, null);
+            XmlNode newExecutionNode = newXmlDoc.CreateNode(XmlNodeType.Element, ExecutionNodeName, null);
 
             if (!parallelTestCount.IsNullOrEmpty())
             {
-                var paralellAttribute = newXmlDoc.CreateAttribute(ParallelTestCountAttributeName);
-                paralellAttribute.Value = parallelTestCount;
-                newExecutionNode.Attributes.Append(paralellAttribute);
+                var parallelAttribute = newXmlDoc.CreateAttribute(ParallelTestCountAttributeName);
+                parallelAttribute.Value = parallelTestCount;
+                newExecutionNode.Attributes?.Append(parallelAttribute);
             }
 
             if (!hostProcessPlatform.IsNullOrEmpty())
             {
                 var hostProcessPlatformAttribute = newXmlDoc.CreateAttribute(HostProcessPlatformAttributeName);
                 hostProcessPlatformAttribute.Value = hostProcessPlatform;
-                newExecutionNode.Attributes.Append(hostProcessPlatformAttribute);
+                newExecutionNode.Attributes?.Append(hostProcessPlatformAttribute);
             }
 
             if (!testTimeout.IsNullOrEmpty())
             {
                 var newTimeoutsNode = newXmlDoc.CreateNode(XmlNodeType.Element, TimeoutsNodeName, null);
-                var testtimeoutattribute = newXmlDoc.CreateAttribute(TestTimeoutAttributeName);
-                testtimeoutattribute.Value = testTimeout;
-                newTimeoutsNode.Attributes.Append(testtimeoutattribute);
+                var testTimeoutAttribute = newXmlDoc.CreateAttribute(TestTimeoutAttributeName);
+                testTimeoutAttribute.Value = testTimeout;
+                newTimeoutsNode.Attributes?.Append(testTimeoutAttribute);
                 newExecutionNode.AppendChild(newXmlDoc.ImportNode(newTimeoutsNode, deep: true));
             }
 
@@ -354,12 +370,17 @@ public class Migrator
     }
 
     /// <summary>
-    /// Adds the datacollector nodes to the runSettings xml.
+    /// Adds the dataCollector nodes to the runSettings xml.
     /// </summary>
-    /// <param name="oldDatacollectorNodes"> Datacollector Nodes</param>
+    /// <param name="oldDataCollectorNodes"> Datacollector Nodes</param>
     /// <param name="newXmlDoc">Xml doc to process</param>
-    private static void AddDataCollectorNodes(XmlNodeList oldDatacollectorNodes, XmlDocument newXmlDoc)
+    private static void AddDataCollectorNodes(XmlNodeList oldDataCollectorNodes, XmlDocument? newXmlDoc)
     {
+        if (newXmlDoc?.DocumentElement == null)
+        {
+            return;
+        }
+
         var dataCollectionRunSettingsNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/DataCollectionRunSettings");
         dataCollectionRunSettingsNode ??= newXmlDoc.CreateNode(XmlNodeType.Element, DataCollectionRunSettingsNodeName, null);
 
@@ -369,12 +390,12 @@ public class Migrator
             dataCollectorsNode = newXmlDoc.CreateNode(XmlNodeType.Element, DataCollectorsNodeName, null);
             dataCollectionRunSettingsNode.AppendChild(newXmlDoc.ImportNode(dataCollectorsNode, deep: true));
             newXmlDoc.DocumentElement.AppendChild(dataCollectionRunSettingsNode);
-            dataCollectorsNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/DataCollectionRunSettings/DataCollectors");
+            dataCollectorsNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/DataCollectionRunSettings/DataCollectors")!;
         }
 
-        foreach (XmlNode datacollector in oldDatacollectorNodes)
+        foreach (XmlNode dataCollector in oldDataCollectorNodes)
         {
-            dataCollectorsNode.AppendChild(newXmlDoc.ImportNode(datacollector, deep: true));
+            dataCollectorsNode.AppendChild(newXmlDoc.ImportNode(dataCollector, deep: true));
         }
     }
 
@@ -385,6 +406,11 @@ public class Migrator
     /// <param name="newXmlDoc">Xml doc to process</param>
     private static void AddRunTimeoutNode(string runTimeout, XmlDocument newXmlDoc)
     {
+        if (newXmlDoc.DocumentElement == null)
+        {
+            return;
+        }
+
         var runConfigurationNode = newXmlDoc.DocumentElement.SelectSingleNode(@"/RunSettings/RunConfiguration");
         runConfigurationNode ??= newXmlDoc.CreateNode(XmlNodeType.Element, RunConfigurationNodeName, null);
 
