@@ -226,7 +226,7 @@ internal class TestPlatform : ITestPlatform
 
         foreach (string source in sources)
         {
-            string sourceDirectory = Path.GetDirectoryName(source);
+            var sourceDirectory = Path.GetDirectoryName(source);
             if (!string.IsNullOrEmpty(sourceDirectory) && _fileHelper.DirectoryExists(sourceDirectory))
             {
                 SearchOption searchOption = GetSearchOption(strategy, SearchOption.TopDirectoryOnly);
@@ -275,7 +275,9 @@ internal class TestPlatform : ITestPlatform
                 .Union(defaultExtensionPaths);
         }
 
-        string extensionsFolder = Path.Combine(Path.GetDirectoryName(typeof(TestPlatform).GetTypeInfo().Assembly.GetAssemblyLocation()), "Extensions");
+        string extensionsFolder = Path.Combine(
+            Path.GetDirectoryName(typeof(TestPlatform).GetTypeInfo().Assembly.GetAssemblyLocation())!,
+            "Extensions");
         if (!fileHelper.DirectoryExists(extensionsFolder))
         {
             // TODO: Since we no-longer run from <playground>\vstest.console\vstest.conosle.exe in Playground, the relative
@@ -283,7 +285,7 @@ internal class TestPlatform : ITestPlatform
             // should come up with a better way of fixing this.
             // NOTE: This is specific to Playground which references vstest.console from a location that doesn't contain
             // the Extensions folder. Normal projects shouldn't have this issue.
-            extensionsFolder = Path.Combine(Path.GetDirectoryName(extensionsFolder), "vstest.console", "Extensions");
+            extensionsFolder = Path.Combine(Path.GetDirectoryName(extensionsFolder)!, "vstest.console", "Extensions");
         }
 
         if (fileHelper.DirectoryExists(extensionsFolder))
