@@ -2,18 +2,14 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-#if !NETSTANDARD1_0
 using System.Collections.Concurrent;
-#endif
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.Reflection;
 using System.Runtime.Serialization;
 
-#if !NETSTANDARD1_0
 using Microsoft.VisualStudio.TestPlatform.Utilities;
-#endif
 
 namespace Microsoft.VisualStudio.TestPlatform.ObjectModel;
 
@@ -22,18 +18,11 @@ public delegate bool ValidateValueCallback(object? value);
 [DataContract]
 public class TestProperty : IEquatable<TestProperty>
 {
-    private Type _valueType;
-#if !NETSTANDARD1_0
     private static readonly ConcurrentDictionary<string, Type> TypeCache = new();
-#else
-    private static readonly Dictionary<string, Type> TypeCache = new();
-#endif
 
-#if NETSTANDARD1_0
-    private static bool DisableFastJson { get; set; } = true;
-#else
     private static bool DisableFastJson { get; set; } = FeatureFlag.Instance.IsSet(FeatureFlag.DISABLE_FASTER_JSON_SERIALIZATION);
-#endif
+
+    private Type _valueType;
 
     //public static Stopwatch
 
@@ -196,11 +185,7 @@ public class TestProperty : IEquatable<TestProperty>
             {
                 if (type != null)
                 {
-#if !NETSTANDARD1_0
                     TypeCache.TryAdd(typeName, type);
-#else
-                    TypeCache[typeName] = type;
-#endif
                     return type;
                 }
             }
@@ -263,11 +248,7 @@ public class TestProperty : IEquatable<TestProperty>
 
         if (!DisableFastJson)
         {
-#if !NETSTANDARD1_0
             TypeCache.TryAdd(typeName, type);
-#else
-            TypeCache[typeName] = type;
-#endif
         }
         return type;
     }
