@@ -11,30 +11,21 @@ namespace Microsoft.VisualStudio.TestPlatform.ObjectModel;
 /// <summary>
 /// Class that holds collection of traits
 /// </summary>
-#if NETFRAMEWORK
 [Serializable]
-#endif
 public class TraitCollection : IEnumerable<Trait>
 {
     internal const string TraitPropertyId = "TestObject.Traits";
     private static readonly TestProperty TraitsProperty = TestProperty.Register(
         TraitPropertyId,
-#if WINDOWS_UWP
-        // TODO: Fix this with proper resourcing for UWP and Win 8.1 Apps
-        // Trying to access resources will throw "MissingManifestResourceException" percolated as "TypeInitialization" exception
-        "Traits",
-#else
+        // REVIEW ME: We should make some test to see if that's causing any crash on WinUI/UWP
         Resources.Resources.TestCasePropertyTraitsLabel,
-#endif
         typeof(KeyValuePair<string, string>[]),
 #pragma warning disable 618
         TestPropertyAttributes.Hidden | TestPropertyAttributes.Trait,
 #pragma warning restore 618
         typeof(TestObject));
 
-#if NETFRAMEWORK
     [NonSerialized]
-#endif
     private readonly TestObject _testObject;
 
     internal TraitCollection(TestObject testObject)
