@@ -33,7 +33,7 @@ public class PathResolver
         {
             var oldFilePath = args[0];
             var newFileName = string.Concat(Path.GetFileNameWithoutExtension(oldFilePath), "_", DateTime.Now.ToString("MM-dd-yyyy_hh-mm-ss", CultureInfo.CurrentCulture), RunSettingsExtension);
-            newFilePath = Path.Combine(Path.GetDirectoryName(oldFilePath), newFileName);
+            newFilePath = Path.Combine(Path.GetDirectoryName(oldFilePath)!, newFileName);
         }
         else if (args.Length == 2)
         {
