@@ -50,8 +50,6 @@ public class IntegrationTestBase
     private readonly string _xUnitTestAdapterRelativePath = @"xunit.runner.visualstudio\{0}\build\_common".Replace('\\', Path.DirectorySeparatorChar);
     private readonly string _chutzpahTestAdapterRelativePath = @"chutzpah\{0}\tools".Replace('\\', Path.DirectorySeparatorChar);
 
-    protected static readonly bool IsWindows = Environment.OSVersion.Platform.ToString().StartsWith("Win");
-
     public enum UnitTestFramework
     {
         NUnit, XUnit, MSTest, CPP, Chutzpah
@@ -626,7 +624,7 @@ public class IntegrationTestBase
         }
         else if (IsNetCoreRunner())
         {
-            var executablePath = IsWindows ? @"dotnet\dotnet.exe" : @"dotnet-linux/dotnet";
+            var executablePath = OSUtils.IsWindows ? @"dotnet\dotnet.exe" : @"dotnet-linux/dotnet";
             consoleRunnerPath = Path.Combine(_testEnvironment.ToolsDirectory, executablePath);
         }
         else
@@ -678,7 +676,7 @@ public class IntegrationTestBase
         var consoleRunnerPath = IsNetCoreRunner()
                 ? GetDotnetRunnerPath()
                 : GetConsoleRunnerPath();
-        var executablePath = IsWindows ? @"dotnet\dotnet.exe" : @"dotnet-linux/dotnet";
+        var executablePath = OSUtils.IsWindows ? @"dotnet\dotnet.exe" : @"dotnet-linux/dotnet";
         var dotnetPath = Path.Combine(_testEnvironment.ToolsDirectory, executablePath);
 
         if (!File.Exists(dotnetPath))
@@ -770,7 +768,7 @@ public class IntegrationTestBase
 
         environmentVariables["DOTNET_MULTILEVEL_LOOKUP"] = "0";
 
-        var executablePath = IsWindows ? @"dotnet\dotnet.exe" : @"dotnet-linux/dotnet";
+        var executablePath = OSUtils.IsWindows ? @"dotnet\dotnet.exe" : @"dotnet-linux/dotnet";
         var patchedDotnetPath = Path.Combine(_testEnvironment.TestArtifactsDirectory, executablePath);
         ExecuteApplication(patchedDotnetPath, string.Join(" ", command, args), out stdOut, out stdError, out exitCode, environmentVariables);
     }
@@ -954,7 +952,7 @@ public class IntegrationTestBase
             architecture == "X86" ?
             "dotnet_x86" :
             $"dotnet",
-            $"dotnet{(IsWindows ? ".exe" : "")}");
+            $"dotnet{(OSUtils.IsWindows ? ".exe" : "")}");
 
         Assert.IsTrue(File.Exists(path));
 
